* libmutil, libmsip: Removed some warnings when compiling to .dll using MSVC++.
git-svn-id: svn://svn.minisip.org/minisip/trunk@1055 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmsip/source/SipUtils.cxx
@@ -43,7 +43,7 @@ bool SipUtils::startsWith(std::string line, std::string part){
 }
 
 int SipUtils::findEndOfHeader(const string &buf, int &startIndex){
-	unsigned bufSize = buf.size();
+	unsigned bufSize = (unsigned)buf.size();
 	int endi=bufSize;
 	int i;
 	int parserState=0;  	// Parser states:
@@ -59,7 +59,7 @@ int SipUtils::findEndOfHeader(const string &buf, int &startIndex){
 				//         return
 
 	i=startIndex;
-	while (i<bufSize && (buf[i]=='\r' || buf[i]=='\n' || buf[i]==' '))
+	while (i<(int)bufSize && (buf[i]=='\r' || buf[i]=='\n' || buf[i]==' '))
 		i++;
 	startIndex = i;
